Fix compilation warnings in store_rabbitkw and store_rabbitv3

+++ ldms/src/store/store_rabbitv3.c
@@ -370,8 +370,8 @@ static int config(ldmsd_plug_handle_t handle, struct attr_value_list *kwl, struc
 			return EINVAL;
 		}
 	}
-	ovis_log(mylog, OVIS_LINFO,"rabbitv3: config host=%s port=%d vhost=%s user=%s metainterval=%d\n",
-	       value, amqp_port, vhvalue, user, g_metainterval);
+	ovis_log(mylog, OVIS_LINFO,"rabbitv3: config host=%s port=%d vhost=%s user=%s metainterval=%ld\n",
+                 value, amqp_port, vhvalue, user, (long)g_metainterval);
 
 #ifdef USEFILTER
 	char *fvalue;
@@ -893,7 +893,7 @@ update_metrics(struct rabbitv3_store_instance *si, ldms_set_t set,
 {
 	if (si->metric_count) {
 		ovis_log(mylog, OVIS_LINFO,"rabbitv3: unexpected "
-		       "reconfigure to size %d from %d.\n",
+		       "reconfigure to size %zd from %d.\n",
 		       metric_count,si->metric_count);
 	}
 	struct rabbitv3_metric_store *ms;
